cwe-tests/python/requirements.txt
# CWE test dependencies — intentionally vulnerable versions
flask==2.3.0
requests==2.28.0
pycryptodome==3.18.0
lxml==4.9.2
pyyaml==5.4.1
